+++ script/eslint/rules/utils.js
@@ -1,0 +1,4 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const experimental_utils_1 = require("@typescript-eslint/experimental-utils");
+exports.createRule = experimental_utils_1.ESLintUtils.RuleCreator(() => "");