+++ src/utilities/nodeTest.ts
@@ -46,7 +46,7 @@ export function shouldAddDollar(node: Node, state: CompilerState): boolean {
         return false;
     }
 
-    if (isFunctionLike(node, state.typeChecker) && node.parent && ts.isImportSpecifier(node.parent)) {
+    if (isFunctionLike(node, state.typeChecker) && !isVariable(node, state.typeChecker)) {
         return false;
     }
 
@@ -115,6 +115,15 @@ export function shouldUseArray(format: ts.ListFormat) {
     return arrayBracketsMap[format] || '';
 }
 
+/**
+ * e.g. let a = []; let b = a;  ==>  $a = array(); $b = &$a;
+ * @param node node
+ */
+export function shouldUseReference(node: ts.Node, typeChecker: ts.TypeChecker) {
+    const nodeType = typeChecker.getTypeAtLocation(node);
+    return ts.isIdentifier(node) && !isFunctionLike(node, typeChecker) && (nodeType.flags === ts.TypeFlags.Object);
+}
+
 const stringLikeType = new Set([
     ts.TypeFlags.String,
     ts.TypeFlags.StringLiteral
@@ -166,6 +175,11 @@ export function isClassLike(node: ts.Node, typeChecker: ts.TypeChecker) {
 export function isClassInstance(node: ts.Node, typeChecker: ts.TypeChecker) {
     const nodeType = typeChecker.getTypeAtLocation(node);
     const nodeSymbol = typeChecker.getSymbolAtLocation(node);
+
+    const baseTypeName = getBaseTypeName(nodeType)
+    if (baseTypeName === 'PHPClass') return true
+    if (baseTypeName === 'PHPArray') return false
+
     if (!nodeSymbol) {
         return false;
     }
@@ -173,10 +187,42 @@ export function isClassInstance(node: ts.Node, typeChecker: ts.TypeChecker) {
         || (nodeType.objectFlags & ts.ObjectFlags.Interface) && nodeType.symbol.getEscapedName() === 'Date';
 }
 
+// 通过最近的父类来确定是用对象（->）还是数组（[]）
+// 优化：遍历父类有的方法来确定用哪个父类
+function getBaseTypeName(nodeType: ts.Type): 'PHPClass' | 'PHPArray' | 'other' {
+    const queue = [nodeType]
+
+    while(queue.length) {
+        const nodeType = queue.shift()
+        const baseTypes = nodeType.getBaseTypes()
+        if (!baseTypes) continue
+
+        for (const baseType of baseTypes) {
+            const baseSymbol = baseType.getSymbol()
+            if (!baseSymbol) continue
+
+            const name = baseSymbol.getName()
+            if (name === 'PHPClass' || name === 'PHPArray') return name
+            else queue.push(baseType)
+        }
+    }
+    return 'other'
+}
+
 export function isFunctionLike(node: ts.Node, typeChecker: ts.TypeChecker) {
     const nodeType = typeChecker.getTypeAtLocation(node);
     const nodeSymbol = nodeType.getSymbol();
     return !!nodeSymbol && nodeSymbol.getFlags() === ts.SymbolFlags.Function;
+}
+
+/**
+ * used for:
+ * variable whos value is function-like but not function declaration
+ */
+export function isVariable(node: ts.Node, typeChecker: ts.TypeChecker) {
+    const nodeType = typeChecker.getTypeAtLocation(node);
+    const nodeSymbol = nodeType.getSymbol();
+    return nodeSymbol.declarations[0].parent.kind === ts.SyntaxKind.VariableDeclaration;
 }
 
 export function isVisibilityModifier(node: ts.Modifier) {