+++ test/features/import.php
@@ -1,13 +1,13 @@
 <?php
 namespace test\case_import;
 require_once(dirname(__FILE__) . '/' . "../some-utils.php");
-use \Other_Utils as Util;
-use \Some_Utils;
+use \someModule\Other_Utils as Util;
+use \someModule\Some_Utils;
 $tplData = array();
 $tplData["src"] = Some_Utils::makeTcLink("url");
 $tplData["title"] = Some_Utils::highlight("title");
 $tplData["title"] = Util::$sample;
-$tplData["title"] = \func() . "aa";
+$tplData["title"] = \someModule\func() . "aa";
 $a = array(
     "test" => "hello"
 );