+++ test/features/stringReplace.ts
@@ -1,0 +1,7 @@
+let a = 'abc';
+
+a.replace(/ab/i, '');
+a.replace(/ab/gi, '');
+a.replace(/ab/ig, '');
+a.replace(/ab/g, '');
+a.replace(/ab/, '');