+++ test/runtime/spec/Ts2Php_HelperTest.php
@@ -47,6 +47,23 @@ final class Ts2Php_HelperTest extends TestCase {
             Ts2Php_Helper::str_slice('abcdefghi', 5),
             'fghi'
         );
+
+        $this->assertEquals(
+            Ts2Php_Helper::str_slice('test一下中文', 3, 6),
+            't一下'
+        );
+    }
+
+    public function testStrSplit() {
+        $this->assertEquals(
+            Ts2Php_Helper::strSplit('', 'abcde'),
+            array('a', 'b', 'c', 'd', 'e')
+        );
+
+        $this->assertEquals(
+            Ts2Php_Helper::strSplit('c', 'abcde'),
+            array('ab', 'de')
+        );
     }
 
     public function testStartsWith() {